webapp/requirements.txt
aniso8601==9.0.1
click==8.1.2
Flask==2.1.1
Flask-Cors==3.0.10
Flask-RESTful==0.3.9
gseapy==0.10.8
h5py==3.6.0
itsdangerous==2.1.2
Jinja2==3.1.1
MarkupSafe==2.1.1
numpy==1.21.6
pandas==1.4.2
plotly==5.7.0
pyOpenSSL==22.0.0
python-dateutil==2.8.2
pytz==2022.1
PyYAML==6.0
scipy==1.8.0
six==1.16.0
Werkzeug==2.1.1
